# static/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/core/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreAuth"]}},"/api/core/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreAuth"]}},"/api/core/organization/create":{"post":{"operationId":"OrganizationController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/info/{organizationId}":{"post":{"operationId":"OrganizationController_info","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/info/edit":{"post":{"operationId":"OrganizationController_infoEdit","parameters":[],"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/member/invite":{"post":{"operationId":"OrganizationController_memberInvite","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationInviteDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/member/dimission":{"post":{"operationId":"OrganizationController_memberDimission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationDimissionDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/members/onjob/{organizationId}":{"post":{"operationId":"OrganizationController_members","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/members/dimissioned/{organizationId}":{"post":{"operationId":"OrganizationController_membersDimissioned","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/organization/members/invited/{organizationId}":{"post":{"operationId":"OrganizationController_membersInvited","parameters":[{"name":"organizationId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["CoreOrganization"]}},"/api/core/team/create":{"post":{"operationId":"TeamController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreTeam"]}},"/api/core/team/info/{teamId}":{"post":{"operationId":"TeamController_info","parameters":[{"name":"teamId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["CoreTeam"]}},"/api/core/team/info/edit":{"post":{"operationId":"TeamController_infoEdit","parameters":[],"responses":{"201":{"description":""}},"tags":["CoreTeam"]}},"/api/core/team/member/add":{"post":{"operationId":"TeamController_memberAdd","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamMemberAddDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreTeam"]}},"/api/core/team/member/remove":{"post":{"operationId":"TeamController_memberRemove","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamMemberRemoveDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreTeam"]}},"/api/core/team/members/{teamId}":{"post":{"operationId":"TeamController_members","parameters":[{"name":"teamId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"tags":["CoreTeam"]}},"/api/core/user/info":{"post":{"operationId":"UserController_info","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommonUserDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreUser"]}},"/api/core/user/info/edit":{"post":{"operationId":"UserController_infoEdit","parameters":[],"responses":{"201":{"description":""}},"tags":["CoreUser"]}},"/api/core/user/organization/invite/pass":{"post":{"operationId":"UserController_organizationInvitePass","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommonOrganizationDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreUser"]}},"/api/core/user/organization/invite/refuse":{"post":{"operationId":"UserController_organizationInviteRefuse","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommonOrganizationDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreUser"]}},"/api/core/user/organizations":{"post":{"operationId":"UserController_organizations","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommonOrganizationDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreUser"]}},"/api/core/user/teams":{"post":{"operationId":"UserController_teams","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommonOrganizationDto"}}}},"responses":{"201":{"description":""}},"tags":["CoreUser"]}},"/api/business/defect/list":{"post":{"operationId":"DefectController_list","parameters":[],"responses":{"201":{"description":""}},"tags":["BusinessDefect"]}},"/api/business/defect/create":{"post":{"operationId":"DefectController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["BusinessDefect"]}},"/api/business/defect/edit":{"post":{"operationId":"DefectController_edit","parameters":[],"responses":{"201":{"description":""}},"tags":["BusinessDefect"]}},"/api/business/defect/detail/{defectId}":{"post":{"operationId":"DefectController_detail","parameters":[],"responses":{"201":{"description":""}},"tags":["BusinessDefect"]}},"/api/business/defect/detail/content/{defectContentId}":{"post":{"operationId":"DefectController_defectContentDetail","parameters":[],"responses":{"201":{"description":""}},"tags":["BusinessDefect"]}}},"info":{"title":"细胞管理系统","description":"细胞、项目管理、任务排期","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthRegisterDto":{"type":"object","properties":{"account":{"type":"string","description":"账号"},"password":{"type":"string","description":"密码"}},"required":["account","password"]},"AuthLoginDto":{"type":"object","properties":{"account":{"type":"string","description":"账号"},"password":{"type":"string","description":"密码"}},"required":["account","password"]},"OrganizationCreateDto":{"type":"object","properties":{"name":{"type":"string","description":"名称"},"user_id":{"type":"number","description":"所有者id"}},"required":["name","user_id"]},"OrganizationInviteDto":{"type":"object","properties":{"organization_id":{"type":"number","description":"组织id"},"user_id":{"type":"number","description":"用户id"}},"required":["organization_id","user_id"]},"OrganizationDimissionDto":{"type":"object","properties":{"organization_id":{"type":"number","description":"组织id"},"user_id":{"type":"number","description":"用户id"}},"required":["organization_id","user_id"]},"TeamCreateDto":{"type":"object","properties":{"name":{"type":"string","description":"名称"},"user_id":{"type":"number","description":"所有者id"},"organization_id":{"type":"number","description":"组织id"}},"required":["name","user_id","organization_id"]},"TeamMemberAddDto":{"type":"object","properties":{"team_id":{"type":"number","description":"团队id"},"user_id":{"type":"number","description":"所有者id"}},"required":["team_id","user_id"]},"TeamMemberRemoveDto":{"type":"object","properties":{"team_id":{"type":"number","description":"团队id"},"user_id":{"type":"number","description":"所有者id"}},"required":["team_id","user_id"]},"CommonUserDto":{"type":"object","properties":{"user_id":{"type":"number","description":"所有者id"}},"required":["user_id"]},"CommonOrganizationDto":{"type":"object","properties":{"user_id":{"type":"number","description":"所有者id"},"organization_id":{"type":"number","description":"组织id"}},"required":["user_id","organization_id"]}}}}